% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{compute_fitted_values_tv}
\alias{compute_fitted_values_tv}
\title{Compute the fitted values of a fully observed process}
\usage{
compute_fitted_values_tv(observed, lambda1_Y, nu_Y, phi_Y, kappa_Y)
}
\arguments{
\item{observed}{The observed sequence of counts}

\item{lambda1_Y}{initial value for \eqn{lambda}}

\item{nu_Y, phi_Y, kappa_Y}{the time-varying parameters of the model
 (\code{kappa} is also allowed to be time-varying here);
vectors of equal length}
}
\description{
Compute the fitted values of a fully observed process, which can also serve as an
approximation to those of an underreported process which is approximated by the given
fully observed process.
}
