% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{compute_sop_tv}
\alias{compute_sop_tv}
\title{Marginal second-order properties of an underreported model}
\usage{
compute_sop_tv(lambda1, nu, phi, kappa, psi, q, compute_Sigma = FALSE)
}
\arguments{
\item{lambda1}{the initial value of lambda}

\item{nu, phi, kappa, psi}{the time-varying parameters of the model
 (\code{kappa} and \code{psi} are also allowed to be time-varying here);
vectors of equal length}

\item{q}{the reporting probability}

\item{compute_Sigma}{logical: shall a full covariance matrix be computed?}
}
\value{
a named list containing the means, variances, first-order autocovariances.
decay parameters of the autocovariance functions and, if desired, full autocovariance
matrix for both \eqn{X} and \eqn{\tilde{X}}.
}
\description{
Compute the time-varying mean and covariance structure of an underreported model.
}
