% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{fit_hhh4u_covariate}
\alias{fit_hhh4u_covariate}
\title{Fit an endemic-epidemic model with underreporting and one covariate}
\usage{
fit_hhh4u_covariate(Y, q, m1, vl1, covariate, initial = c(alpha_nu = 4,
  beta_nu = 0, alpha_phi = -1, alpha_kappa = -1, log_psi = -3),
  max_lag = 10, iter_optim = 3, ...)
}
\arguments{
\item{Y}{a time series of counts (numeric vector)}

\item{q}{the assumed reporting probability}

\item{m1}{the initial mean, i.e. $E(lambda_1)$}

\item{vl1}{the initial variance of lambda, i.e. $Var(lambda_1)$}

\item{covariate}{the values of the covariate entering into the endemic component (numeric vector)}

\item{initial}{the initial value of the parameter vector passed to optim
(note: the function tries different starting values in any case)}

\item{max_lag}{in evaluation of likelihood only lags up to max_lag are taken into account}
}
\value{
the return object from \code{optim} providing the maximum likelihood estimates
(mostly on the log scale).
}
\description{
Fits an endemic-epidemic model with underreporting and one covariate in the endemic component
using an approximative maximum likelihood scheme. The likelihood approximation is based on an
approximation of the process by a second-order equivalent process with complete reporting.
The reporting probability cannot be estimated from the data (in most cases these contain no
information on it) and thus needs to be specified in advance.
}
