% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneous.R
\name{fit_hhh4u_homog}
\alias{fit_hhh4u_homog}
\title{Fitting time-homogeneous underreported endemic-epidemic model using maximum likelihood}
\usage{
fit_hhh4u_homog(observed, q, include_kappa = TRUE, initial = c(log_nu =
  2, log_phi = -1, log_kappa = -2, log_psi = -3), max_lag = 5,
  iter_optim = 3, hessian = TRUE, ...)
}
\arguments{
\item{observed}{a time series of counts (numeric vector)}

\item{q}{the assumed reporting probability}

\item{initial}{the initial value of the parameter vector passed to optim
(note: the function re-runs optimization several times)}

\item{max_lag}{in evaluation of (conditional) likelihood only lags up to \code{max_lag} are taken into account}
}
\value{
a named list containing the following elements:
\itemize{
  \item parameter estimates on the original scale (code{par}) and associated
  standard errors (\code{se}), obtained via the delta method
  \item parameter estimates on the log scale (as used in the optimization; \code{par_log}),
  associated standard errors (\code{se_log}) and inverse Fisher information matrix (\code{Sigma_log})
  \item the return object of the call to \code{optim} (\code{opt})
  \item the data (\code{observed}), reporting probability (\code{q}) and other settings \code{settings}
  used in the call to \code{fit_hhh4u_homog}
}
}
\description{
Fits a time-homogeneous endemic-epidemic model with underreporting using an approximative maximum
likelihood scheme. The likelihood approximation is based on an approximation of the process by
a second-order equivalent process with complete reporting. The reporting probability cannot be
estimated from the data (in most cases these contain no information on it) and thus needs to be
specified in advance.
}
