% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{fit_hhh4u_seasonal}
\alias{fit_hhh4u_seasonal}
\title{Fits a sesaonal endemic-epidemic model with underreporting using an approximative maximum
likelihood scheme. The likelihood approximation is based on an approximation of the process by
a second-order equivalent process with complete reporting. The reporting probability cannot be
estimated from the data (in most cases these contain no information on it) and thus needs to be
specified in advance.}
\usage{
fit_hhh4u_seasonal(observed, L, q, seas_phi = FALSE,
  include_kappa = TRUE, initial = c(alpha_nu = 4, gamma_nu = 0,
  delta_nu = 0, alpha_phi = -1, gamma_phi = 0, delta_phi = 0, alpha_kappa =
  -1, log_psi = -3), max_lag = 10, iter_optim = 3, ...)
}
\arguments{
\item{observed}{a time series of counts (numeric vector)}

\item{L}{the number of observations per period (integer; e.g. 52 for weekly data and yearly seasonality)}

\item{q}{the assumed reporting probability}

\item{seas_phi}{should seasonality also be accounted for in the autoregressive parameter $phi$?
or only in the immigration parameter $nu$?}

\item{include_kappa}{should the parameter $kappa$ and thus lagged versions of the conditional expectation $lambda$ be included?}

\item{initial}{the initial value of the parameter vector passed to optim
(note: the function tries different starting values in any case)}

\item{max_lag}{in evaluation of likelihood only lags up to max_lag are taken into account}
}
\value{
the return object from \code{optim} providing the maximum likelihood estimates
(mostly on the log scale).
}
\description{
Fits a sesaonal endemic-epidemic model with underreporting using an approximative maximum
likelihood scheme. The likelihood approximation is based on an approximation of the process by
a second-order equivalent process with complete reporting. The reporting probability cannot be
estimated from the data (in most cases these contain no information on it) and thus needs to be
specified in advance.
}
