% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags.R
\name{fit_jags}
\alias{fit_jags}
\title{Bayesian fitting of a time-homogeneous underreported endemic-epidemic model using JAGS}
\usage{
fit_jags(Y, p, n.chains = 5, n.iter = 10000, n.burnin = 1000,
  n.thin = 10, ...)
}
\arguments{
\item{Y}{a time series of counts (numeric vector)}

\item{p}{the assumed reporting probability}

\item{n.chains, n.iter, n.burnin, n.thin}{settings passed to R2jags::jags}
}
\description{
Fits a time-homogeneous endemic-epidemic model with underreporting using JAGS.
}
