% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{hhh4u}
\alias{hhh4u}
\title{Fit an endemic-epidemic model to underreported data}
\usage{
hhh4u(stsObj, control = list(ar = list(f = ~1, use_kappa = TRUE), end =
  list(f = ~1), family = c("Poisson", "NegBin1"), q = NULL, subset =
  1:nrow(stsObj), data = list(t = stsObj@epoch - min(stsObj@epoch)), start
  = NULL, optimizer = list(maxit = 5000), return_se = TRUE))
}
\arguments{
\item{stsObj}{an sts object (see documentation of the \code{surveillance} package)
containing the observed time series. Only univariate time series are supported.}

\item{control}{a control list, similar to the one used in \code{hhh4} from the \code{surveillance} package.
The following aspects can be specified:
\itemize{
\item \code{ar}: List specifying the autoregressive/epidemic component. Possible elements:
\itemize{
\item \code{f}: a formula for \eqn{\log(\phi_t)}, using e.g. \code{surveillance::addSeason2formula}
to account for seasonality
\item \code{use_kappa}: should a feedback mechanism, ie autoregression on \eqn{lambda_{t - 1}} be used?
If \code{FALSE} the parameter \eqn{\kappa} is forced to zero.
}
\item \code{end}: List specifying the endemic component. Possible elements:
\itemize{
\item \code{f}: a formula for \eqn{\log(\nu)}
}
\item \code{family} The distibutional family, either \code{"Poisson"} or \code{"NegBin1"} for a negative binomial.
\item \code{q} The assumed reporting probability
\item \code{subset} The subset of the data to which the model shall be fitted; needs to contained within \code{1:nrow(stsObj@observed) }
\item \code{start} Initial value passed to \code{optim}; needs to have the correct length and naming.
\item \code{optimize} Additional arguments passed to \code{optim}
\item \code{return_se} Should the standard errors be returned? In case of numerical problems it can be reasonable to deactivate this.
}}
}
\value{
An object of class \code{hhh4u}, which is a named list with the following elements:
\itemize{
\item \code{coefficients}: The parameter estimates
\item \code{se}: The estimated standard errors of the coefficients
\item \code{cov}: The estimated autocovariance matrix of the parameter estimates
\item \code{par_long}: a list containing the time-varying parameters in vector form
\item \code{par_long_approximation}: a list containing the time-varying parameters of the approximating fully observed process
\item \code{fitted.values}: the fitted values
\item \code{dim}: the model dimension (i.e. number of parameters)
\item \code{loglikelihood:} the log-likelihood of the model
\item \code{convergence}: code for convergence status returned by \code{optim}; 0 indicates successful convergence.
\item \code{control}: the control list used for the fit
\item \code{stsObj}: the \code{sts} object used for the fit
\item \code{nobs}: the number of observations used for the fit
\item \code{optim}: the return object of the call to optim
}
}
\description{
Fits a time-homogeneous endemic-epidemic model with underreporting using an approximative maximum
likelihood scheme.
}
\details{
This function can be used in a similar way as \code{surveillance::hhh4}, which fits endemic-epidemic
models without accounting for underreporting. However, unlike \code{surveillance::hhh4}, \code{hhh4u}
is limited to the analysis of univariate time series.

Denote by \eqn{X_t} the number of cases of a given disease in a given population in week \eqn{t}.
The endemic-epidemic model is defined as
\deqn{X_{t} | past \sim NegBin(\lambda_t, \psi)}
\deqn{\lambda_t = \nu_t + \phi_t X_{t - 1} + \kappa\lambda_{t - 1}.}
Here, the negative binomial distribution is parameterized via the (conditional) mean \eqn{\lambda_t}
and an overdispersion parameter \eqn{\psi} so that \eqn{Var(X_t) | \lambda_t = \lambda_t + \psi\lambda_t^2}.
\eqn{\nu_t} is denoted as the endemic component, the two autoregressive terms \eqn{\phi X_{t - 1}}
and \eqn{\kappa\lambda_{t - 1}} are called the epidemic component.
The time-varying parameters \eqn{\nu} and \eqn{\phi} are modelled in a log-linear way, e.g.
\deqn{\log(\nu) = \alpha_\nu + \gamma_\nu \sin(2\pi t/52) + \delta_\nu \cos(2\pi t/52)}
to model yearly seasonality in weekly data.

We here assume that the data are undereported, i.e. we do not observe the \eqn{X_t}, but \eqn{X'_t} which
are linked to \eqn{X_t} through independent binomial thinning:
\deqn{X'_t | X \sim Bin(X_t, q),}
where \eqn{q} is the reporting probability. The reporting probability cannot be
estimated from the data (in most cases these contain no information on it) and thus needs to be
specified in advance.

The likelihood approximation is based on an approximation of the process by a second-order
equivalent process with complete reporting.

For details see Bracher/Held: A marginal moment matching approach for fiting endemic-epidemic models
to underreported disease surveillance counts.
}
