% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{llik_cov}
\alias{llik_cov}
\title{Evaluate the approximate log-likelihood of an underreported endemic-epidemic
model with time-varying parameters}
\usage{
llik_cov(Y, m1, vl1, nu, phi, kappa, psi, q, max_lag = 5,
  return_contributions = FALSE)
}
\arguments{
\item{Y}{a time series of counts (numeric vector)}

\item{m1}{the initial mean, i.e. $E(lambda_1)$}

\item{vl1}{the initial variance of lambda, i.e. $Var(lambda_1)$}

\item{nu, phi, kappa, psi}{the time-varying model parameters (vectors of same length)}

\item{psi}{overdispersion parameter (scalar)}

\item{q}{the assumed reporting probability}

\item{max_lag}{in evaluation of likelihood only lags up to max_lag are taken into account}

\item{return_contributions}{shall the log-likelihood contributions of each time point be
returned (as vector)?}
}
\value{
The log-likelihood as scalar or (if \code{return_contributions == TRUE}) the vector of
log-likelihood contributions.
}
\description{
The likelihood approximation is based on an approximation of the process by
a second-order equivalent process with complete reporting. Note that this is
an R version while a reimplementation in Rcpp is used in the optimization.
}
