% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{llik_tv}
\alias{llik_tv}
\title{Evaluate the (approximate) likelihood of an underreported model}
\usage{
llik_tv(observed, lambda1, nu, phi, kappa, psi, q,
  return_contributions = FALSE, log = TRUE)
}
\arguments{
\item{observed}{The observed sequence of counts}

\item{lambda1}{initial value for \eqn{lambda}}

\item{nu, phi, kappa, psi}{the time-varying parameters of the model
 (\code{kappa} is also allowed to be time-varying here);
vectors of equal length}

\item{q}{the assumed reporting probability (potentially time-varying)}

\item{return_contributions}{logical: should the likelihood contributions
of the different weeks be returned?}

\item{log}{logical: return log-likelihood?}
}
\description{
Compute the fitted values of a fully observed process, which can also serve as an
approximation to those of an underreported process which is approximated by the given
fully observed process.
}
