% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{marg_mean_tv}
\alias{marg_mean_tv}
\title{Marginal mean of an underreported model}
\usage{
marg_mean_tv(lambda1, nu, phi, kappa, q = 1)
}
\arguments{
\item{lambda1}{the initial value of lambda}

\item{nu, phi, kappa}{the time-varying parameters of the model (\code{kappa} is also allowed to be time-varying here);
vectors of equal length}

\item{q}{the potentially time-varying reporting probability; defaults to 1.}
}
\value{
a vector containing the marginal means
}
\description{
Compute the time-varying marginal means of an underreported model.
}
