% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneous.R
\name{nllik_homog}
\alias{nllik_homog}
\title{Evaluate the approximate log-likelihood of a time-homogeneous underreported endemic-epidemic model}
\usage{
nllik_homog(
  observed,
  nu,
  phi,
  kappa,
  psi,
  q,
  max_lag = 5,
  return_contributions = FALSE
)
}
\arguments{
\item{observed}{a time series of counts (numeric vector)}

\item{nu, phi, kappa, psi}{the model parameters (scalars)}

\item{q}{the assumed reporting probability}

\item{max_lag}{in evaluation of likelihood only lags up to max_lag are taken into account}

\item{return_contributions}{shall the log-likelihood contributions of each time point be
returned (as vector)?}
}
\value{
The log-likelihood as scalar or (if \code{return_contributions == TRUE}) the vector of
log-likelihood contributions.
}
\description{
The likelihood approximation is based on an approximation of the process by
a second-order equivalent process with complete reporting.
}
