% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{nllik_seas}
\alias{nllik_seas}
\title{Evaluate the approximate negative log-likelihood of a seasonal underreported endemic-epidemic model}
\usage{
nllik_seas(observed, alpha_nu, gamma_nu, delta_nu, alpha_phi,
  gamma_phi = 0, delta_phi = 0, alpha_kappa, psi, q, L, max_lag = 5,
  return_contributions = FALSE)
}
\arguments{
\item{observed}{a time series of counts (numeric vector)}

\item{alpha_nu, gamma_nu, delta_nu}{the endemic model parameters (scalar)}

\item{alpha_phi, gamma_phi, delta_phi, alpha_kappa}{the autoregressive model parameters (scalars)}

\item{psi}{overdispersion parameter (scalar)}

\item{q}{the assumed reporting probability}

\item{L}{the period length (integer; e.g. 52 for weekly data and yearly seasonality)}

\item{max_lag}{in evaluation of likelihood only lags up to max_lag are taken into account}

\item{return_contributions}{shall the log-likelihood contributions of each time point be
returned (as vector)?}
}
\value{
The negative log-likelihood as scalar or (if \code{return_contributions == TRUE}) the vector of
negative log-likelihood contributions.
}
\description{
The likelihood approximation is based on an approximation of the process by
a second-order equivalent process with complete reporting.
}
