% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{reparam_tv}
\alias{reparam_tv}
\title{Approximate an underreported model by a fully observed one}
\usage{
reparam_tv(lambda1, nu, phi, kappa, psi, q)
}
\arguments{
\item{nu, phi, kappa, psi}{the time-varying parameters of the model
 (\code{kappa} and \code{psi} are also allowed to be time-varying here);
vectors of equal length}

\item{q}{the reporting probability}
}
\description{
This function returns the parameters of a fully observed model with the same second-order properties
as the specified underreported model.
}
