% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{rota_germany}
\alias{rota_germany}
\title{Data set on rotavirus gastroenteritis in Germany}
\source{
Surveillance counts retrieved from SurvStat@RKI 2.0 service (https://survstat.rki.de), Robert Koch Institute, Berlin as of 17 Oct 2018.

Surveillance counts retrieved from SurvStat@RKI 2.0 service (https://survstat.rki.de), Robert Koch Institute, Berlin as of 17 Oct 2018.
}
\description{
Case counts of rotavirus gastroenteritis in the sixteen German states, 2001-2017

Case counts of norovirus gastroenteritis in the sixteen German states, 2001-2017
}
\author{
Johannes Bracher

Johannes Bracher
}
\keyword{data}
