% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{setControl}
\alias{setControl}
\title{Check and complete the control list}
\usage{
setControl(control, stsObj)
}
\arguments{
\item{control}{the control argument as provided by the user}

\item{stsObj}{the stsObj to be used for the fit}
}
\value{
a complete control list, with default values added where nothing was specified.
}
\description{
Similar to \code{surveillance:::setControl and surveillance:::makeControl} this function
checks the arguments in the control list and adds default settings where necessary.
}
