% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{simulate_hhh4u}
\alias{simulate_hhh4u}
\title{Simulate from underreported model with time-varying parameters}
\usage{
simulate_hhh4u(
  lambda1,
  nu,
  phi,
  kappa,
  psi,
  q = 1,
  decoarsen = FALSE,
  lgt = max(sapply(list(nu, phi, kappa, psi, q), length)),
  return_sts = TRUE,
  args_sts = list()
)
}
\arguments{
\item{lambda1}{the initial value of lambda}

\item{nu, phi, kappa, psi}{the model parameters (vectors of equal length
corresponding to the length of the time series; scalars are recycled).}

\item{q}{the reporting probability; either a scalar or a vector of length lgt}

\item{lgt}{= leng}
}
\value{
A named list with elements \code{"X"}, \code{"X_tilde"} and \code{"lambda"} containing
the unthinned and thinned simulated time series as well as the conditional mean process.
}
\description{
Simulate from a time-homogeneous endemic-epidemic model with underreporting.
Can also be used through the wrapper \code{simulate_hhh4u}.
}
\examples{
sim <- simulate_hhh4u(lambda1 = 1, nu = c(1:5, 4:1), phi = 0.4, kappa = 0.2, psi = 0.1, q = 0.5)
}
