% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{simulate_hhh4u2}
\alias{simulate_hhh4u2}
\title{Simulate from underreported model with time-varying parameters}
\usage{
simulate_hhh4u2(
  lambda1,
  beta_nu,
  beta_phi,
  kappa,
  psi,
  q = 1,
  lgt,
  z_nu = matrix(rep(1, lgt)),
  z_phi = matrix(rep(1, lgt))
)
}
\arguments{
\item{lambda1}{the initial value of lambda}

\item{beta_nu}{parameters for the linear form entering into \eqn{log(nu)}}

\item{beta_phi}{parameters for the linear form entering into \eqn{log(phi)}}

\item{kappa, psi}{Time-constant model parameters}

\item{q}{the reporting probability}

\item{z_nu}{Matrix containing the covariates entering into \eqn{log(nu)} (in columns)}

\item{z_phi}{Matrix containing the covariates entering into \eqn{log(phi)} (in columns)}
}
\value{
A named list with elements \code{"X"}, \code{"X_tilde"} and \code{"lambda"} containing
the unthinned and thinned simulated time series as well as the conditional mean process.
}
\description{
Simulate from a time-homogeneous endemic-epidemic model with underreporting.
Wrapper around \code{simulate_hhh4u0} which takes vector-valued parameters.
}
\examples{
beta_nu <- c(1, 0.005)
z_nu <- cbind(1, 1:100)
sim <- simulate_hhh4u2(lambda1 = 1, beta_nu = beta_nu, beta_phi = -0.6,
kappa = 0.2, psi = 0.1, q = 0.5, z_nu = z_nu)
}
