% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\name{simulate_hhh4u_covariate}
\alias{simulate_hhh4u_covariate}
\title{Simulate from underreported model with one covariate}
\usage{
simulate_hhh4u_covariate(lambda1, alpha_nu, beta_nu, phi, kappa, psi, q = 1, z)
}
\arguments{
\item{lambda1}{the initial mean}

\item{alpha_nu, beta_nu, phi, kappa, psi}{the parameters of the latent model}

\item{q}{the reporting probability}

\item{z}{a covariate entering into the endemic parameter $nu$}
}
\value{
A named list with elements \code{"X"} and \code{"Y"} containing the unthinned and thinned simulated time series.
}
\description{
Simulate from an endemic-epidemic model with a covariate entering
into the endemic component. Requires specification of the initial value for $lambda$.
}
\examples{
# define a covariate:
z <- sin(2*pi*1:200/20)
# define the model parameters:
lambda1 <- 10
alpha_nu <- 3
beta_nu <- 1
phi <- 0.5
kappa <- 0.2
psi <- 0.1
q <- 0.5
sim <- generate_ar_cov(lambda1 = lambda1,
                       alpha_nu = alpha_nu,
                       beta_nu = beta_nu,
                       phi = phi,
                       kappa = kappa,
                       psi = psi,
                       q = q,
                       z = z)
}
