% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{simulate_hhh4u_seasonal}
\alias{simulate_hhh4u_seasonal}
\title{Simulate from seasonal underreported model}
\usage{
simulate_hhh4u_seasonal(alpha_nu, gamma_nu, delta_nu, alpha_phi, gamma_phi,
  delta_phi, kappa, psi, q = 1, L = 52, n_seas = 10, start = 10,
  burn_in = 10)
}
\arguments{
\item{alpha_nu, gamma_nu, delta_nu, alpha_phi, gamma_phi, delta_phi, kappa, psi}{the model parameters}

\item{q}{the reporting probability}

\item{L}{the length of one cycle (integer)}

\item{n_seas}{the number of seasons to simulate (integer)}

\item{start}{initial value of both $X$ and $lambda$ (at beginning of burn in period; can usually be ignored)}

\item{burn_in}{number of seasons to discard to reach stationary phase}
}
\value{
A named list with elements \code{"X"} and \code{"Y"} containing the unthinned and thinned simulated time series.
}
\description{
Simulate from a seasonal endemic-epidemic model with underreporting. Includes a burn-in period
to reach stationary phase.
}
